chill-requirements.txt
nibabel
numpy
geodistk
macholib
matplotlib
nibabel
opencv-python
pyinstaller
pyqt6
pyside6
qt-material
qt6-tools
scikit-image
simplecrf
scikit-learn
torchaudio
torchvision
wincertstore